+++ .changeset/free-areas-rule.md
@@ -1,0 +1,5 @@
+---
+'@orfium/ictinus': minor
+---
+
+badge and tag group
