--- apps/storybook/CHANGELOG.md
@@ -1,87 +1,0 @@
-# @app/storybook
-
-## 0.0.10
-
-### Patch Changes
-
-- Updated dependencies [cc639ea]
-  - @orfium/ictinus@5.41.4
-
-## 0.0.9
-
-### Patch Changes
-
-- Updated dependencies [d6aba99]
-  - @orfium/ictinus@5.41.3
-
-## 0.0.8
-
-### Patch Changes
-
-- 4f8756d: fix table nested header borders
-- b3660fc: text fixes
-- Updated dependencies [4f8756d]
-- Updated dependencies [b3660fc]
-  - @orfium/ictinus@5.41.2
-
-## 0.0.7
-
-### Patch Changes
-
-- Updated dependencies [5ed22bd]
-  - @orfium/ictinus@5.41.1
-
-## 0.0.6
-
-### Patch Changes
-
-- 68a6d1e: new box
-- Updated dependencies [c3bd67e]
-- Updated dependencies [68a6d1e]
-  - @orfium/ictinus@5.41.0
-  - @orfium/tokens@5.1.2
-
-## 0.0.5
-
-### Patch Changes
-
-- c39c33a: chore: resolve any issues with exportation of libraries and how storybook consumes ictinus for DEV mode
-- Updated dependencies [c39c33a]
-  - @orfium/ictinus@5.40.1
-  - @orfium/tokens@5.1.1
-
-## 0.0.4
-
-### Patch Changes
-
-- ff25264: add code variant for tag
-- 4d88250: chore: split tokens from ictinus
-
-  Split the token library from ictinus.
-  Ensure the library can export CSS tokens.
-  Create examples demonstrating the use of the CSS tokens.
-
-- Updated dependencies [ff25264]
-- Updated dependencies [4d88250]
-  - @orfium/ictinus@5.40.0
-  - @orfium/tokens@5.1.0
-
-## 0.0.3
-
-### Patch Changes
-
-- Updated dependencies [d31fbb8]
-  - @orfium/ictinus@5.39.5
-
-## 0.0.2
-
-### Patch Changes
-
-- Updated dependencies [e8c779e]
-  - @orfium/ictinus@5.39.4
-
-## 0.0.1
-
-### Patch Changes
-
-- 9ceafe6: changing repo to monorepo
