+++ apps/storybook/package.json
@@ -1,7 +1,7 @@
 {
   "name": "storybook",
   "private": true,
-  "version": "0.0.10",
+  "version": "0.0.0",
   "type": "module",
   "scripts": {
     "dev": "storybook dev -p 6006",
@@ -28,6 +28,7 @@
     "polished": "^4.3.1",
     "react-docgen-typescript": "^2.4.0",
     "react-router": "^7.12.0",
+    "react-stately": "^3.41.0",
     "remark-gfm": "^4.0.1"
   },
   "devDependencies": {
