+++ packages/ictinus/src/tag-group/index.ts
@@ -1,0 +1,1 @@
+export * from './TagGroup';
