--- packages/ictinus/src/vanilla/Tag/index.ts
@@ -1,1 +1,0 @@
-export * from './Tag';
