+++ packages/ictinus/src/vanilla/index.ts
@@ -10,13 +10,13 @@ export * from '../layers';
 export * from '../nav';
 export * from '../skeleton';
 export * from '../sprinkles';
+export * from '../tag-group';
 export * from '../utils/cn';
 export * from './Box';
 export * from './Dialog';
 export * from './Menu';
 export * from './Popover';
 export * from './Table';
-export * from './Tag';
 export * from './Text';
 export * from './ThemeProvider';
 export * from './Tooltip';
